Fix date bug
When using the `YYYY-MM-DD` format, the month and day had to be
zero-padded otherwise it wouldn't be a valid date. Decided it would be
safer to switch to `MM/DD/YYYY` format which allows for non-zero pad.
Remembering to zero pad is tough so this seems like a safer approach.