Add Other Weeks nav list to mobile post page
Adds a button at the top of the post page that lets you navigate to
other posts. Only appears on smaller screens, since on large screens
you can use the sidenav.

While implementing this, I found the transition animation to be
distracting. I don't like that it results in the sidenav animating. So
I removed that for now.