Generate standings table automatically
Replaces the manual process of updating the standings table in the
contentful post with a custom contentful entry that will be replaced
with a table that fetches the standings from supabase.

Includes several changes to make this work:
- Setup the supabase client
- Create an API route to fetch the standings for a given week range
- Create a contentful entry for the standings that is replaced by html
  - Ideally we could use a svelte component to replace this instead of
  raw html but one thing at a time

Also did some rearranging of the various models and utils to hopefully
allow us to scale our file structure better.

And switched to pnpm.